requirements.txt
# LiveKit Core
livekit>=1.0.20
livekit-agents>=1.3.6
livekit-api>=1.1.0

# LiveKit Plugins
livekit-plugins-google>=1.3.6
livekit-plugins-elevenlabs>=1.3.6
livekit-plugins-soniox>=1.3.6
livekit-plugins-deepgram>=1.3.6
livekit-plugins-openai>=1.3.6
livekit-plugins-silero>=1.3.6
# cartesia not used - removed

# Core Dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
pydantic>=2.0.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

